edited

+++ src/main/java/bank/Account.java
@@ -11,7 +11,6 @@ public class Account {
     setBalance(balance);
   }
 
-
   public int getId() {
     return this.Id;
   }
@@ -28,12 +27,12 @@ public class Account {
     this.type = type;
   }
 
-  public int getBalance() {
+  public double getBalance() {
     return this.balance;
   }
 
-  public void setBalance(int balance) {
+  public void setBalance(double balance) {
     this.balance = balance;
   }
-
+  
 }
